add graphql server
