+++ applications/tari_indexer/Cargo.toml
@@ -38,6 +38,8 @@ tari_validator_node_client = { path = "../../clients/validator_node_client" }
 anyhow = "1.0.53"
 async-trait = "0.1.50"
 axum = "0.6.0"
+async-graphql = "5.0.7"
+async-graphql-axum = "5.0.7"
 axum-jrpc = { version = "0.3.2", features = ["anyhow_error"] }
 chrono = "0.4.22"
 ciborium = { version = "0.2.0" }
