+++ clients/tari_indexer_client/Cargo.toml
@@ -12,6 +12,8 @@ tari_common_types = { git = "https://github.com/tari-project/tari.git", tag = "v
 tari_comms_logging = { path = "../../comms/tari_comms_logging" }
 
 anyhow = "1.0.65"
+async-graphql-axum = "5.0.7"
+juniper = "0.15.11"
 reqwest = { version = "0.11.11", features = ["json"] }
 serde = "1.0.144"
 serde_json = "1.0.85"
